requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# Core ML and NLP libraries
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
chromadb>=0.4.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Environment and configuration
python-dotenv>=1.0.0

# Document processing
pypdf>=3.17.0

# Email functionality
secure-smtplib>=0.1.1

# Utility libraries
tqdm>=4.64.0
requests>=2.28.0

# Development tools
pytest>=7.4.0
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0

# Logging and monitoring
loguru>=0.7.0

# Optional: For enhanced performance
# faiss-cpu>=1.7.4  # Uncomment for better vector search performance
